+++ src/main/java/com/dkthanh/demo/domain/dto/OptionDto.java
@@ -16,8 +16,8 @@ public class OptionDto {
     private Integer optionId;
     private String optionName;
     private String optionDescription;
-    private Integer fundMin;
+    private Long fundMin;
     private List<ItemEntity> items;
     private Integer projectId;
-    private Integer pledge;
+    private Long pledge;
 }
